… src/test/regress/sql/vacuum.sql