Beginning in PG12 https://github.com/postgres/postgres/commit/e0c2933a7, vacuumdb issues a separate VACUUM command for each targeted table. This change resulted in a performance regression in upstream due to spending O(N^2) time in vac_update_datfrozenxid(). 

PG16 https://github.com/postgres/postgres/commit/a46a701 introduced options to control whether VACUUM runs vac_update_datfrozenxid. This PR backports the upstream commit.

Issuing separate VACUUM's for each table has an even greater performance regression in GPDB due to the dispatch overhead. Instead we refactor how vacuumdb builds VACUUM commands.

When vacuumdb is run without the --jobs/--tables flags, execute a single database wide VACUUM. This restores the behavior of 6X vacuumdb.

When vacuumdb is run with --jobs/--tables, divide the table list evenly among available connections and issue one VACUUM per connection.

Performance on a DB with 30k tables:

main branch:
```
vacuumdb postgres  0.28s user 0.20s system 0% cpu 8:29.80 total
vacuumdb postgres --jobs 2  0.36s user 0.30s system 0% cpu 6:35.34 total
vacuumdb postgres --jobs 8  0.45s user 0.57s system 0% cpu 2:58.57 total
```
PR:
```
vacuumdb postgres  0.01s user 0.00s system 0% cpu 3:39.39 total
vacuumdb postgres --jobs 2  0.00s user 0.01s system 0% cpu 2:24.89 total
vacuumdb postgres --jobs 8  0.01s user 0.00s system 0% cpu 1:37.36 total
```