@ashwinstar 2023-08-14
- Understand GPDB has more impact without and hence also more gain with this PR, though impact must be seen on PostgreSQL as well?
- Would be nice to have more explanation on the batch-size functionality. Also, would love to hear from you on its relevance to upstream as well?