@soumyadeep2007 2023-08-16
For the upper limit for the batch size, there is no real upper limit on the number of relations that can be specified in a VACUUM command. Thinking more about it, it is a bad experience for users to come and run the VACUUM with a high batch size, only for it to error out with the VMEM error that you were describing to me today.

We can probably cap it at an arbitrary limit, based on the default VMEM settings perhaps.