@bmdoil 2023-08-17
> For the upper limit for the batch size, there is no real upper limit on the number of relations that can be specified in a VACUUM command. Thinking more about it, it is a bad experience for users to come and run the VACUUM with a high batch size, only for it to error out with the VMEM error that you were describing to me today.
> 
> We can probably cap it at an arbitrary limit, based on the default VMEM settings perhaps.

I looked into the VMEM issue more. I had `optimizer=on` and `optimizer_analyze_root_partition=on` and it fell over in `leaf_parts_analyzed` on the coordinator. This is a known issue that was addressed in https://github.com/greenplum-db/gpdb/pull/10499 but it seems I've hit an upper limit. 

```
(gdb) bt
#0  0x00005592e69023a4 in _bt_first (scan=0x5592ea1fb0e0, dir=ForwardScanDirection) at nbtsearch.c:922
#1  0x00005592e68ff51f in btgettuple (scan=0x5592ea1fb0e0, dir=ForwardScanDirection) at nbtree.c:327
#2  0x00005592e68f57db in index_getnext_tid (scan=0x5592ea1fb0e0, direction=ForwardScanDirection) at indexam.c:526
#3  0x00005592e68f5962 in index_getnext_slot (scan=0x5592ea1fb0e0, direction=ForwardScanDirection, slot=0x5592ea1f8cc0) at indexam.c:618
#4  0x00005592e68f4623 in systable_getnext (sysscan=0x5592eb091100) at genam.c:450
#5  0x00005592e6b755ee in fetch_leaf_att_stats (leafRelid=39861, leafAttNum=1) at analyzeutils.c:1138
#6  0x00005592e6b75959 in leaf_parts_analyzed (attrelid=26103, relid_exclude=26487, va_cols=0x5592f9651860, elevel=13) at analyzeutils.c:1285
#7  0x00005592e6b6c096 in expand_vacuum_rel (vrel=0x5592ea25a4c0, options=1) at vacuum.c:1045
#8  0x00005592e6b6b02d in vacuum (relations=0x5592e9ec3488, params=0x7fff76ee9510, bstrategy=0x5592ea1f4d60, isTopLevel=true) at vacuum.c:385
#9  0x00005592e6b6ad8d in ExecVacuum (pstate=0x5592ea1192c0, vacstmt=0x5592ea4a5f50, isTopLevel=true, auto_stats=false) at vacuum.c:252
```
**I'm a little surprised that this is being called since I didn't specify `vacuumdb --analyze`.** 

Per the comment in `expand_vacuum_rel` 
```
/*
* GPDB: If you explicitly ANALYZE a partition, also update the
* parent's stats after the partition has been ANALYZEd. (Thanks to
* the code to merge leaf statistics, it should be fast.)
*
```
We don't skip partition children and if optimizer_analyze_root_partition=on, we're descending into `leaf_parts_analyzed`. Is this a bug?
```
		if (optimizer_analyze_root_partition && !skip_this)
		{
                                .......
				/*
				 * Only ANALYZE the parent if the stats can be updated by merging
				 * child stats.
				 */
				if (!leaf_parts_analyzed(parent_relid, child_relid, vrel->va_cols, elevel))
					break;
```
VACUUM without a relation list is fine because we don't call `expand_vacuum_rel` 

After disabling `optimizer_analyze_root_partition` I had no issues running a VACUUM with a list of 30,000 tables.

With this context, `batch-size` feels like a workaround for a bug. I don't think we need to add the option at all.