@ashwinstar 2023-08-17
> We don't skip partition children and if optimizer_analyze_root_partition=on, we're descending into leaf_parts_analyzed. Is this a bug?
Maybe this discussion is relevant https://github.com/greenplum-db/gpdb/pull/16218 for this @chrishajas