@soumyadeep2007 2023-08-18
> We don't skip partition children and if optimizer_analyze_root_partition=on, we're descending into leaf_parts_analyzed. Is this a bug?

Wow this is a terrible bug. There is no reason for us to do this check `leaf_parts_analyzed` or add to the vacuum list if `analyze` has not even been specified.

We should guard like this: `if ((options & VACOPT_ANALYZE) && optimizer_analyze_root_partition && !skip_this)` and then that should solve this problem.

> After disabling optimizer_analyze_root_partition I had no issues running a VACUUM with a list of 30,000 tables.

> With this context, batch-size feels like a workaround for a bug. I don't think we need to add the option at all.

I agree.